requirements.txt
langchain[all]
google-search-results
python-dotenv
streamlit-chat